+++ src/Connection/AuthenticationException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Disque\Connection;
+
+class AuthenticationException extends ConnectionException
+{
+}