+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 All Notable changes will be documented in this file. This project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## [1.2.2]
+## [1.3.0] - 2015-05-18
 
 ### Added
 - Added support for `WORKING`.
 - Added `processing()` method to Queue API.
+- Added `$password` option to `addServer()` in `Disque\Client`.
+
+### Changed
+- By default when creating a new `Disque\Client` without arguments NO server
+is pre-loaded. You will have to manually add servers via `addServer()`, or
+specify them to the `Disque\Client` constructor.
 
 ## [1.2.1] - 2015-05-14
 
@@ -94,7 +100,8 @@ parameters were specified, an `InvalidCommandArgumentException` was thrown.
 - Added support for Predis connections, and allowing adding new connection
 methods via `ConnectionInterface`.
 
-[1.2.2]: https://github.com/mariano/disque-php/compare/1.2.1...HEAD
+[1.3.1]: https://github.com/mariano/disque-php/compare/1.3.1...HEAD
+[1.3.0]: https://github.com/mariano/disque-php/compare/tag/1.3.0
 [1.2.1]: https://github.com/mariano/disque-php/releases/tag/1.2.1
 [1.2.0]: https://github.com/mariano/disque-php/releases/tag/1.2.0
 [1.1.0]: https://github.com/mariano/disque-php/releases/tag/1.1.0